[REFACTOR] 뷰 명칭 변경
