+++ frontend/src/views/main/MainView.vue
@@ -5,7 +5,7 @@ import { RouterLink } from "vue-router";
 <template>
   <div class="text-center">
     <h1>Welcome to My App</h1>
-    <RouterLink to="/projects" class="btn btn-primary">Go to MyProjects</RouterLink>
+    <RouterLink to="/projects" class="btn btn-primary">MY PROJECT LIST</RouterLink>
 
     <RouterLink to="/projectBoards" class="btn btn-primary">PROJECT BOARD LIST</RouterLink>
   </div>
